pyproject.toml
[tool.poetry]
name = "konbinein"
version = "0.0.0"
description = ""
authors = ["kinoubenkyou <kinoubenkyou@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.1"
django = "^4.2.8"
psycopg2 = "^2.9.5"
djangorestframework = "^3.14.0"
factory-boy = "^3.2.1"
pycountry = "^22.3.5"
celery = "^5.3.4"
redis = "^5.0.1"
drf-spectacular = "^0.26.5"
mongoengine = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
preview = true

[tool.isort]
profile = "black"
